+++ apps/web/src/main.tsx
@@ -1,10 +1,5 @@
-import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "@atlaskit/css-reset";
-import { App } from "./App.js";
+import { Root } from "./Root.js";
 
-createRoot(document.getElementById("root")!).render(
-  <StrictMode>
-    <App />
-  </StrictMode>
-);
+createRoot(document.getElementById("root")!).render(<Root />);